fix(similarity-prep): align module imports after split
